Create PrimitiveTypesChallenge.java
Challenge: 
Create a byte variable and set it to any valid byte number. Create a short variable and set it to any valid short number. Create an int variable and set it to any valid int number. Lastly, create a variable of type long and make it equal to 50000 plus 10 times the sum of the byte plus the short plus the integer values.